+++ Models/TouchEvent.cs
@@ -12,8 +12,8 @@ namespace CPRTouchVision.Models
         public float Y { get; set; }
         public float NormalizedX { get; set; }
         public float NormalizedY { get; set; }
-        public ushort GameScreenX { get; set; }
-        public ushort GameScreenY { get; set; }
+        public float GameScreenX { get; set; }
+        public float GameScreenY { get; set; }
         public int Id { get; set; }
         public float Radius { get; set; }
         //public DateTime Timestamp { get; set; }
